+++ src/invoices/Voucher/VoucherDetails.test.js
@@ -27,4 +27,17 @@ describe('VoucherDetails component', () => {
     expect(getByText(testVoucher.status)).toBeDefined();
     expect(getByText('ui-invoice.invoice.enclosureNeeded')).toBeDefined();
   });
+
+  it('should display No account in voucher details', () => {
+    const { getByText } = renderVoucherDetails(testVoucher);
+
+    expect(getByText('ui-invoice.invoice.details.voucher.noAccount')).toBeDefined();
+  });
+
+  it('should display account number in voucher details', () => {
+    const accountNo = 'accountNo';
+    const { getByText } = renderVoucherDetails({ ...testVoucher, accountNo });
+
+    expect(getByText(accountNo)).toBeDefined();
+  });
 });